Refactor `Cryptol.Symbolic.Prims` and `Cryptol.Symbolic.Value` into a single new module `Crypol.Eval.SBV`.
